+++ examples/pyslop_extensions/detect_shims/__init__.py
@@ -1,0 +1,1 @@
+"""detect-shims extension package."""